CPG-207 added test cases for model classes (Session and User)
